# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Category API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/menu_inventory": {
            "get": {
                "tags": [
                    "MenuInventory"
                ],
                "summary": "Get all menu inventory items",
                "description": "Retrieve a list of all menu inventory items",
                "operationId": "2d5149373d20f9e16795f02a6364af01",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "query",
                        "description": "Restaurant ID to filter inventory",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of menu inventory items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MenuInventory"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No menu inventory items found"
                    }
                }
            },
            "post": {
                "tags": [
                    "MenuInventory"
                ],
                "summary": "Create a new menu inventory item",
                "description": "Add a new item to the menu inventory",
                "operationId": "90dcc163fbb05248a517b2225d09da5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuInventory"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu inventory item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuInventory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/menu_inventory/{id}": {
            "get": {
                "tags": [
                    "MenuInventory"
                ],
                "summary": "Get a specific menu inventory item",
                "description": "Retrieve a specific menu inventory item by ID",
                "operationId": "0c8ec584ca6345e1afda5c29399df8f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu inventory item details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuInventory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu inventory item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "MenuInventory"
                ],
                "summary": "Update a menu inventory item",
                "description": "Update details of a menu inventory item",
                "operationId": "8bae0551969e6bb26705ccb4454e0496",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MenuInventory"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu inventory item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuInventory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Menu inventory item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "MenuInventory"
                ],
                "summary": "Delete a menu inventory item",
                "description": "Delete a menu inventory item by ID",
                "operationId": "a6833f3e1f2c8cf7b6168dc91c200a9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu Inventory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu inventory item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Menu inventory item deleted successfully"
                            }
                        }
                    },
                    "404": {
                        "description": "Menu inventory item not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MenuInventory": {
                "title": "Menu Inventory",
                "description": "Menu Inventory schema",
                "properties": {
                    "id": {
                        "description": "Menu Inventory ID",
                        "type": "integer",
                        "example": 1
                    },
                    "menuId": {
                        "description": "ID of the menu item",
                        "type": "integer",
                        "example": 5
                    },
                    "restaurantId": {
                        "description": "Restaurant ID associated with the inventory",
                        "type": "string",
                        "example": "R123456789"
                    },
                    "quantity": {
                        "description": "Quantity of the menu item",
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "stockId": {
                        "description": "ID of the associated stock item",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-11T13:55:52.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-12T13:55:52.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}